Fix printing struct timeval on FreeBSD
Do not use format that assumes the type of tv_sec and tv_usec,
instead use %jd and cast those fields to intmax_t

+++ tests/ts_print_raw.c
@@ -16,6 +16,7 @@
 #include <sys/ioctl.h>
 #include <sys/mman.h>
 #include <sys/time.h>
+#include <inttypes.h>
 
 #include "tslib.h"
 
@@ -56,7 +57,8 @@ int main()
 		if (ret != 1)
 			continue;
 
-		printf("%ld.%06ld: %6d %6d %6d\n", samp.tv.tv_sec, samp.tv.tv_usec, samp.x, samp.y, samp.pressure);
+		printf("%jd.%06jd: %6d %6d %6d\n", (intmax_t)samp.tv.tv_sec,
+			(intmax_t)samp.tv.tv_usec, samp.x, samp.y, samp.pressure);
 
 	}
 }
